# Lower interpolation into a call to concat