--- a/src/Compiler/Checking/CheckExpressions.fs
+++ b/src/Compiler/Checking/CheckExpressions.fs
@@ -7336,6 +7336,33 @@ and TcInterpolatedStringExpr cenv (overallTy: OverallTy) env m tpenv (parts: Syn
             // Type check the expressions filling the holes
             let fillExprs, tpenv = TcExprsNoFlexes cenv env m tpenv argTys synFillExprs
 
+            // If all fill expressions are strings and there is less then 5 parts of the interpolated string total
+            // then we can use System.String.Concat instead of a sprintf call
+            if isString && (parts.Length < 5) && (argTys |> List.forall (isStringTy g)) then
+                let rec f xs ys acc =
+                    match xs with
+                    | SynInterpolatedStringPart.String(s, m)::xs ->
+                        let sb = System.Text.StringBuilder(s).Replace("%%", "%")
+                        f xs ys ((mkString g m (sb.ToString()))::acc)
+                    | SynInterpolatedStringPart.FillExpr(_, _)::xs ->
+                        match ys with
+                        | y::ys -> f xs ys (y::acc)
+                        | _ -> error(Error((0, "FOOBAR"), m)) // TODO XXX wrong error
+                    | _ -> acc
+
+                let args = f parts fillExprs [] |> List.rev
+                assert (args.Length = parts.Length)
+                if args.Length = 4 then
+                    (mkStaticCall_String_Concat4 g m args[0] args[1] args[2] args[3], tpenv)
+                elif args.Length = 3 then
+                    (mkStaticCall_String_Concat3 g m args[0] args[1] args[2], tpenv)
+                elif args.Length = 2 then
+                    (mkStaticCall_String_Concat2 g m args[0] args[1], tpenv)
+                else
+                    // Throw some error
+                    error(Error((0, "FOOBAR2"), m)) // TODO XXX wrong error
+            else // TODO XXX messed up indentation below
+
             let fillExprsBoxed = (argTys, fillExprs) ||> List.map2 (mkCallBox g m)
 
             let argsExpr = mkArray (g.obj_ty, fillExprsBoxed, m)