--- a/src/Compiler/Utilities/ReadOnlySpan.fs
+++ b/src/Compiler/Utilities/ReadOnlySpan.fs
@@ -59,4 +59,19 @@ type ReadOnlySpanExtensions =
                 i <- i - 1
 
         if found then i else -1
+
+    [<Extension>]
+    static member LastIndexOfAnyExcept(span: ReadOnlySpan<char>, value: char) =
+        let mutable i = span.Length - 1
+        let mutable found = false
+
+        while not found && i >= 0 do
+            let c = span[i]
+
+            if c <> value then
+                found <- true
+            else
+                i <- i - 1
+
+        if found then i else -1
 #endif