… docs/release-notes/.FSharp.Compiler.Service/8.0.300.md
… docs/release-notes/.Language/preview.md
… src/Compiler/FSComp.txt
… src/Compiler/Facilities/LanguageFeatures.fs
… src/Compiler/Utilities/ReadOnlySpan.fsi
… src/Compiler/pars.fsy
… src/Compiler/xlf/FSComp.txt.cs.xlf
… src/Compiler/xlf/FSComp.txt.de.xlf
… src/Compiler/xlf/FSComp.txt.es.xlf
… src/Compiler/xlf/FSComp.txt.fr.xlf
… src/Compiler/xlf/FSComp.txt.it.xlf
… src/Compiler/xlf/FSComp.txt.ja.xlf
… src/Compiler/xlf/FSComp.txt.ko.xlf
… src/Compiler/xlf/FSComp.txt.pl.xlf
… src/Compiler/xlf/FSComp.txt.pt-BR.xlf
… src/Compiler/xlf/FSComp.txt.ru.xlf
… src/Compiler/xlf/FSComp.txt.tr.xlf
… src/Compiler/xlf/FSComp.txt.zh-Hans.xlf
… src/Compiler/xlf/FSComp.txt.zh-Hant.xlf
… tests/AheadOfTime/Trimming/check.ps1
… tests/FSharp.Compiler.ComponentTests/EmittedIL/StringFormatAndInterpolation.fs
… tests/FSharp.Compiler.ComponentTests/Language/InterpolatedStringsTests.fs