## Description

<!-- Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change. -->

Optimization that lowers string interpolation into a call to concat iff there are at most 4 string parts and all fill expressions are strings.

Fixes #16247

## Benchmarks

Run a benchmark like in [this gist](https://gist.github.com/abonie/af32962df723a4a53a4e96ed84eb1f88) with and without the feature flag set.

Results without the flag (**no optimization**):
```

BenchmarkDotNet v0.13.12, Windows 11 (10.0.22621.3155/22H2/2022Update/SunValley2)
11th Gen Intel Core i7-11850H 2.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.300-preview.24101.10
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI DEBUG
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI


```
| Method         | Mean      | Error    | StdDev   | Gen0   | Allocated |
|--------------- |----------:|---------:|---------:|-------:|----------:|
| SingleVariable |  70.49 ns | 1.214 ns | 1.136 ns | 0.0107 |     136 B |
| JustLiterals   | 145.61 ns | 1.443 ns | 1.350 ns | 0.0215 |     272 B |
| Variables      | 144.05 ns | 0.959 ns | 0.897 ns | 0.0215 |     272 B |
| Function       | 116.73 ns | 1.557 ns | 1.381 ns | 0.0191 |     240 B |
| LongString     | 363.35 ns | 2.838 ns | 2.654 ns | 0.4330 |    5432 B |

Results with the flag set (**with optimization**):
```

BenchmarkDotNet v0.13.12, Windows 11 (10.0.22621.3155/22H2/2022Update/SunValley2)
11th Gen Intel Core i7-11850H 2.50GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 8.0.300-preview.24101.10
  [Host]     : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI DEBUG
  DefaultJob : .NET 8.0.2 (8.0.224.6711), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI


```
| Method         | Mean        | Error     | StdDev    | Median      | Gen0   | Allocated |
|--------------- |------------:|----------:|----------:|------------:|-------:|----------:|
| SingleVariable |   0.0004 ns | 0.0013 ns | 0.0011 ns |   0.0000 ns |      - |         - |
| JustLiterals   |  12.0562 ns | 0.0719 ns | 0.0673 ns |  12.0500 ns | 0.0032 |      40 B |
| Variables      |  11.9894 ns | 0.0810 ns | 0.0718 ns |  12.0101 ns | 0.0032 |      40 B |
| Function       |  12.2912 ns | 0.1201 ns | 0.1124 ns |  12.3161 ns | 0.0032 |      40 B |
| LongString     | 206.0906 ns | 2.4311 ns | 2.0301 ns | 206.5404 ns | 0.4158 |    5216 B |
## Checklist

- [x] Test cases added
- [x] Performance benchmarks added in case of performance changes
- [x] Release notes entry updated