@abonie 2024-02-15
Also, if anyone is curious about same benchmarks but **with server GC**:
No optimization:

| Method         | Mean      | Error    | StdDev   | Gen0   | Allocated |
|--------------- |----------:|---------:|---------:|-------:|----------:|
| SingleVariable |  78.08 ns | 1.021 ns | 0.905 ns | 0.0005 |     136 B |
| JustLiterals   | 157.44 ns | 2.675 ns | 2.502 ns | 0.0010 |     272 B |
| Variables      | 151.16 ns | 1.498 ns | 1.401 ns | 0.0010 |     272 B |
| Function       | 125.50 ns | 1.239 ns | 1.099 ns | 0.0010 |     240 B |
| LongString     | 641.34 ns | 2.861 ns | 2.536 ns | 0.0219 |    5432 B |


With optimization:
| Method         | Mean        | Error     | StdDev    | Median      | Gen0   | Allocated |
|--------------- |------------:|----------:|----------:|------------:|-------:|----------:|
| SingleVariable |   0.0064 ns | 0.0103 ns | 0.0096 ns |   0.0000 ns |      - |         - |
| JustLiterals   |  14.8223 ns | 0.1345 ns | 0.1258 ns |  14.8525 ns | 0.0001 |      40 B |
| Variables      |  14.1441 ns | 0.1240 ns | 0.1099 ns |  14.1558 ns | 0.0001 |      40 B |
| Function       |  14.2070 ns | 0.1159 ns | 0.1028 ns |  14.2344 ns | 0.0001 |      40 B |
| LongString     | 508.0675 ns | 6.9382 ns | 6.1506 ns | 507.0075 ns | 0.0210 |    5216 B |